Move helper functions to runtime-util
